Update Auto Entity settings
This commit updates the configuration so that the module fills the field
before the save. This addresses the issue that I was seeing with the title
not showing the auto filled value on the first save. A cache refresh and
redisplay of the page "fixes" that, but it will be confusing for users.